+++ seminars/seminar3/Figure.java
@@ -1,7 +1,7 @@
 package OOP_all.seminars.seminar3;
 
 public abstract class Figure {
-    public abstract double getArea();
+    public abstract int getArea();
 
     public abstract String getName();
